+++ src/main/java/com/ormi/cookudasse/home/controller/HomeController.java
@@ -1,0 +1,23 @@
+package com.ormi.cookudasse.home.controller;
+
+import com.ormi.cookudasse.post.entitiy.Post;
+import com.ormi.cookudasse.post.service.PostService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import java.util.List;
+
+@Controller
+@RequiredArgsConstructor
+public class HomeController {
+    private final PostService postService;
+
+    @GetMapping("/")
+    public String home(Model model) {
+        List<Post> posts = postService.getAllPosts();
+        model.addAttribute("posts", posts);
+        return "home";
+    }
+}
