--- src/main/resources/templates/detailpage/style.css
@@ -1,112 +1,0 @@
-body {
-  font-family: Arial, sans-serif;
-  margin: 0;
-  padding: 0;
-  background-color: #f0f0f0;
-}
-
-.container {
-  max-width: 800px;
-  margin: 0 auto;
-  background-color: white;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-header {
-  text-align: center;
-  padding: 20px 0;
-}
-
-.logo {
-  max-width: 150px;
-}
-
-main {
-  padding: 20px;
-}
-
-.form-group {
-  margin-bottom: 15px;
-}
-
-label {
-  display: block;
-  margin-bottom: 5px;
-}
-
-input[type="text"],
-select,
-textarea {
-  width: 100%;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-textarea {
-  height: 150px;
-}
-
-.button-group {
-  text-align: right;
-}
-
-.btn {
-  padding: 10px 20px;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.btn-primary {
-  background-color: #007bff;
-  color: white;
-}
-
-.btn-secondary {
-  background-color: #6c757d;
-  color: white;
-}
-
-.pagination {
-  text-align: center;
-  padding: 20px 0;
-}
-
-.prev,
-.next {
-  background: none;
-  border: none;
-  font-size: 18px;
-  cursor: pointer;
-}
-
-.post-view h1 {
-  margin-bottom: 10px;
-}
-
-.post-info {
-  color: #666;
-  margin-bottom: 20px;
-}
-
-.post-content h2 {
-  margin-top: 20px;
-}
-
-.reaction-buttons {
-  margin-top: 20px;
-}
-
-.reaction-btn {
-  font-size: 24px;
-  background: none;
-  border: none;
-  cursor: pointer;
-  margin-right: 10px;
-}
-
-.action-buttons {
-  margin-top: 20px;
-  text-align: right;
-}
